requirements.txt
streamlit>=1.28
groq
python-dotenv
pyyaml
